requirements.txt
PyHamcrest==2.0.2
pytest==5.3.5
requests==2.23.0
selenium==3.141.0

